Compliance update: Switched final evaluation score to strict binary integer (0 or 1)

+++ schemas.py
@@ -20,7 +20,7 @@ class StepInfo(BaseModel):
     token: str
     action: str
     is_sensitive: bool
-    final_score: Optional[float] = None
+    final_score: Optional[int] = None
 
 class StepResult(BaseModel):
     observation: Observation
